requirements.txt
ultralytics
streamlit
pandas
numpy
plotly
PILLOW
opencv-python-headless